move utils to correct location
