+++ src/app/static/src/app/types.ts
